requirements.txt
snowflake-connector-python


